requirements-gui.txt
-r requirements-web.txt
pywebview==6.1
